+++ spec/spec_helper.rb
@@ -4,6 +4,7 @@ SPEC_ROOT = File.expand_path(File.dirname(__FILE__)) # ENV["RAILS_ENV"] ||= 'tes
 # require 'rspec/rails'
 
 require_relative '../lib/bluff.rb'
+require_relative 'support/bluffable_class.rb'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.