+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Jan 11 2012 - v0.0.3
+
+* Added architecture for multiple backend/orm support (only activerecord provided so far)
+* Fixed bug with bluff extensions not locating existing classes correctly
+
 ## Jan 11 2012
 
 * Added initial implementation with a small set of built-in data bluffs